Rename sandbox registration file
